refactored navbar from main into new navbarManager class
