+++ ReaderAndroid/app/src/main/java/com/kontinua/readerandroid/MainActivity.kt
@@ -33,6 +33,8 @@ import androidx.core.view.GestureDetectorCompat
 import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import com.google.android.material.navigation.NavigationView
+import com.kontinua.readerandroid.NavbarManager.ChapterData
+import com.kontinua.readerandroid.NavbarManager.WorkbookData
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
@@ -51,8 +53,7 @@ import retrofit2.http.Url
 
 class MainActivity :
     AppCompatActivity(),
-    GestureDetector.OnGestureListener,
-    NavigationView.OnNavigationItemSelectedListener {
+    GestureDetector.OnGestureListener {
 
     private lateinit var imageView: ImageView
     private lateinit var loadingTextView: TextView
@@ -71,9 +72,6 @@ class MainActivity :
     private val baseUrl = "http://10.0.2.2:8000/"
     private var pdfFileName = "workbook-01.pdf"
     private var metaFileName = "workbook-01.json"
-    private val chapterMap = mutableMapOf<Int, Int>()
-    private val workbookMap = mutableMapOf<Int, WorkbookData>()
-    private var workbookSelected = true
     private lateinit var annotationView: AnnotationView
 
     // timer variables
@@ -88,6 +86,8 @@ class MainActivity :
     private var timeLeftMillis: Long = timerDuration
     private var isTimerRunning: Boolean = false // state management
     private var elapsedTimeMillis: Long = 0
+
+    private lateinit var navbarManager: NavbarManager
 
     interface ApiService {
         @GET
@@ -172,15 +172,19 @@ class MainActivity :
         // Retrofit setup moved here
         apiService = retrofit().create(ApiService::class.java)
         loadPdfFromUrl()
-        loadChaptersFromServer()
+
+        // Initialize SidebarManager
+        navbarManager = NavbarManager(this, drawerLayout, chapterView, workbookView, apiService)
+        navbarManager.setupNavbar(toolbar)
+        navbarManager.loadChaptersFromServer(baseUrl, metaFileName)
 
         openWorkbookNavButton.setOnClickListener {
-            if (workbookSelected) {
-                loadWorkbooks()
-                workbookSelected = false
+            if (navbarManager.getWorkbookSelected()) {
+                navbarManager.loadWorkbooks()
+                navbarManager.setWorkbookSelected(false)
             } else {
-                loadChaptersFromServer()
-                workbookSelected = true
+                navbarManager.loadChaptersFromServer(baseUrl, metaFileName)
+                navbarManager.setWorkbookSelected(true)
             }
         }
 
@@ -191,47 +195,10 @@ class MainActivity :
         nextButton.setOnClickListener {
             goToNextPage()
         }
-
-        val toggle = ActionBarDrawerToggle(
-            this,
-            drawerLayout,
-            toolbar,
-            R.string.navigation_drawer_open,
-            R.string.navigation_drawer_close,
-        )
-        drawerLayout.addDrawerListener(toggle)
-        toggle.syncState()
-
-        chapterView.setNavigationItemSelectedListener(this)
-        workbookView.setNavigationItemSelectedListener(this)
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.toolbar_menu, menu)
-        return true
-    }
-
-    override fun onNavigationItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            else -> {
-                if (workbookSelected) {
-                    Log.d("MainActivity", "Clicked on Chapter: ${item.title}")
-                    currentPageIndex = chapterMap[item.itemId]!!
-                    annotationView.setPage(currentPageIndex - 1)
-                    displayPage(currentPageIndex - 1)
-                } else {
-                    Log.d("MainActivity", "Clicked on Workbook: ${item.title}")
-                    pdfFileName = workbookMap[item.itemId]!!.pdfName
-                    metaFileName = workbookMap[item.itemId]!!.metaName
-                    currentPageIndex = 0
-                    workbookSelected = true
-                    loadChaptersFromServer()
-                    loadPdfFromUrl()
-                    annotationView.setWorkbook(pdfFileName)
-                }
-            }
-        }
-        drawerLayout.closeDrawer(GravityCompat.START)
         return true
     }
 
@@ -558,8 +525,8 @@ class MainActivity :
         CoroutineScope(Dispatchers.Main).launch {
             val renderer = pdfRenderer ?: return@launch
             if (pageNumber >= 0 && pageNumber < renderer.pageCount) {
+                annotationView.setPage(pageNumber)
                 displayPage(pageNumber)
-                annotationView.setPage(currentPageIndex)
             } else {
                 // Handle invalid page number (e.g., show an error message)
                 Log.e("MainActivity", "Invalid page number entered")
@@ -587,60 +554,6 @@ class MainActivity :
     @SuppressLint("SetTextI18n")
     private fun updatePageNumberEditText() {
         pageNumberEditText.setText((currentPageIndex + 1).toString())
-    }
-
-    data class ChapterData(val title: String, val id: String, val chap_num: Int, val start_page: Int)
-    data class WorkbookData(val id: String, val metaName: String, val pdfName: String)
-
-    private fun populateMenu(chapters: List<ChapterData>) {
-        val menu = chapterView.menu
-        menu.clear() // Clear existing items
-
-        for (chapter in chapters) {
-            val menuItem = menu.add(
-                Menu.NONE,
-                chapter.id.hashCode(),
-                Menu.NONE,
-                "${chapter.chap_num}. ${chapter.title}",
-            )
-            menuItem.setShowAsAction(MenuItem.SHOW_AS_ACTION_NEVER) // Ensure it stays in the sidebar
-            chapterMap[chapter.id.hashCode()] = chapter.start_page
-        }
-    }
-
-    private fun loadChaptersFromServer() {
-        CoroutineScope(Dispatchers.IO).launch {
-            try {
-                val response = apiService.getChapters("$baseUrl/meta/$metaFileName").execute()
-
-                if (response.isSuccessful) {
-                    val chapters = response.body()
-                    Log.d("MainActivity", "Server Response: $chapters")
-
-                    withContext(Dispatchers.Main) {
-                        if (chapters != null) {
-                            populateMenu(chapters)
-                        } else {
-                            Log.e("MainActivity", "Chapters data is NULL")
-                        }
-                    }
-                } else {
-                    Log.e("MainActivity", "Failed to fetch chapters: ${response.errorBody()?.string()}")
-                }
-            } catch (e: Exception) {
-                Log.e("MainActivity", "Network Error: ${e.message}")
-            }
-        }
-    }
-
-    private fun populateWorkbookMenu(workbooks: List<WorkbookData>) {
-        val menu = chapterView.menu
-        menu.clear()
-
-        for (workbook in workbooks) {
-            val menuItem = menu.add(Menu.NONE, workbook.id.hashCode(), Menu.NONE, workbook.id)
-            workbookMap[workbook.id.hashCode()] = workbook
-        }
     }
 
     private fun startTimer() {
@@ -713,20 +626,17 @@ class MainActivity :
         timerFillView.layoutParams = params
     }
 
-    private fun loadWorkbooks() {
-        CoroutineScope(Dispatchers.IO).launch {
-            val response = apiService.getWorkbooks().execute()
+    fun loadPageNumber(pageNumber: Int){
+        currentPageIndex = pageNumber
+        annotationView.setPage(pageNumber-1)
+        displayPage(pageNumber-1)
+    }
 
-            if (response.isSuccessful) {
-                val workbooks = response.body()
-                if (workbooks != null) {
-                    withContext(Dispatchers.Main) {
-                        populateWorkbookMenu(workbooks)
-                    }
-                }
-            } else {
-                Log.e("MainActivity", "Failed to load workbooks: ${response.errorBody()?.string()}")
-            }
-        }
+    fun loadNewWorkbook(workbook: WorkbookData) {
+        pdfFileName = workbook.pdfName
+        metaFileName = workbook.metaName
+        navbarManager.loadChaptersFromServer(baseUrl, metaFileName)
+        loadPdfFromUrl()
+        annotationView.setWorkbook(pdfFileName)
     }
 }
